+++ face-autocapture-with-liveness-detection/server/config/demoLogConf.js
@@ -41,47 +41,80 @@ const LOG_FILE_NAME_REGEX = (() => {
     return undefined;
 })();
 
+/** @type {InspectOptions} */
+const FORMAT_OPTIONS = {
+    depth: 5, // it will print 5 levels of the object tree in logs (default 2)
+    maxArrayLength: 300 // default limit is 100, increased a bit to log properly arrays
+};
+
 const asyncLocalStorage = new AsyncLocalStorage();
 
+// The logger instance that will be exported
 const logger = createLogger(); // create only one instance shared between all files
 
 /**
- * Create one instance of logger object
+ * Create an instance of logger object
  * @returns {winston.Logger}
  */
 function createLogger() {
-    const rawLogFormat = config.LOG_FORMAT === 'raw';
+    // The message transform function (raw or json)
+    const transform = config.LOG_FORMAT === 'raw' ? transformToRaw : transformToJson;
     const logger = winston.createLogger({
         level: getLogLevel(config.LOG_LEVEL),
         format: format.combine(
             format(info => {
                 return !LOG_FILE_NAME_REGEX || LOG_FILE_NAME_REGEX.test(getFileName()) ? info : null;
             })(),
-            format.timestamp({ format: 'YYYY-MM-DD HH:mm:ss.SSS' }),
+            // In raw, use local time to display technical logs, in json use ISO8601 with timezone offset
+            format.timestamp({ format: config.LOG_FORMAT === 'raw' ? 'YYYY-MM-DD HH:mm:ss.SSS' : 'YYYY-MM-DDTHH:mm:ss.SSSZ' }),
             format.printf(info => {
-                const fileName = getFileName();
+                info.class = getFileName();
                 // Extract contextual tenantId / sessionId
                 const store = asyncLocalStorage.getStore();
-                const tenantId = store?.tenantId || '';
-                const sessionId = store?.sessionId || '';
+                info.tenantId = store?.tenantId || '';
+                info.sessionId = store?.sessionId || '';
                 // Now add header to original message
-                const message = rawLogFormat
-                    ? `${info.timestamp}|${info.thread}|${tenantId}|${sessionId}|${colorize(info.level, info.level.toUpperCase())}|${fileName}|${info.message}`
-                    : {
-                        timestamp: info.timestamp,
-                        thread: info.thread.toString(),
-                        tenantId,
-                        sessionId,
-                        level: info.level.toUpperCase(),
-                        class: fileName,
-                        message: info.message
-                    };
-
-                return rawLogFormat ? message : JSON.stringify(message);
+                return transform(info);
             })
         ),
         defaultMeta: { thread: process.pid.toString().padStart(14, '0') } // to fit native format
     });
+    configure(logger);
+    return logger;
+}
+
+/**
+ * Transform the info object into a raw string
+ * @param {object} info the input object to transform
+ * @param {(level: string, message: string) => string} colorize a function used to colorize text
+ * @return {string} a raw string containing transformed info
+ */
+function transformToRaw(info, addColor = colorize) {
+    return `${info.timestamp}|${info.thread}|${info.tenantId}|${info.sessionId}|${addColor(info.level, info.level.toUpperCase())}|${info.class}|${info.message}`;
+}
+
+/**
+ * Transform the info object into a JSON string
+ * @param {object} info the input object to transform
+ * @returns {string} a JSON string containing transformed info
+ */
+function transformToJson(info) {
+    return JSON.stringify({
+        timestamp: info.timestamp,
+        thread: info.thread.toString(),
+        tenantId: info.tenantId,
+        sessionId: info.sessionId,
+        level: info.level.toUpperCase(),
+        class: info.class,
+        message: info.message
+    });
+}
+
+/**
+ * Configure logger
+ * @param {winston.Logger} logger
+ */
+function configure(logger) {
     if (config.LOG_APPENDER !== 'console') {
         logger.add(new winston.transports.File({
             filename: `${config.LOG_FILE_PATH}/bioserver-demo-error.log`,
@@ -91,9 +124,10 @@ function createLogger() {
     } else {
         logger.add(new winston.transports.Console());
     }
+
     /**
      * Update the current context of this logger
-     * @param {object} context contains the tenantId and sessionId to store
+     * @param {object?} context contains the tenantId and sessionId to store
      */
     logger.updateContext = function (context) {
         const store = asyncLocalStorage.getStore();
@@ -111,20 +145,21 @@ function createLogger() {
             }
         }
     };
+
     // Override log methods ('debug', 'info'...) to use util.format to have similar behaviour as console.log or debug lib
     Object.keys(logger.levels).forEach(level => {
         const logMethod = logger[level];
         logger[level] = function (...args) {
-            logMethod(util.format(...args));
+            logMethod(util.formatWithOptions(FORMAT_OPTIONS, ...args));
         };
     });
+
     // Override 'log' as well
     const _log = logger.log;
     logger.log = function (level, ...args) {
         // Need to attach the logger to 'this' otherwise an error is thrown
-        _log.call(logger, level, util.format(...args));
+        _log.call(logger, level, util.formatWithOptions(FORMAT_OPTIONS, ...args));
     };
-    return logger;
 }
 
 /**
@@ -182,6 +217,9 @@ function getLogLevel(logLevel) {
 module.exports.getLogger = () => {
     return logger;
 };
+
+// For json to raw converter
+module.exports.transformToRaw = transformToRaw;
 
 // For unit tests
 module.exports.createLogger = createLogger;