+++ ReleaseNotes.md
@@ -2,6 +2,13 @@
 This demo app showcases the integration of the IDEMIA WebCapture SDK, including both Face Autocapture & Liveness and Identity Document Autocapture features.
 
 ---
+## October 6, 2025 : Face Autocapture and Face Liveness
+
+### What's New
+* Reduce zoom on face displayed in the result screen.
+* Rework of the result screen for passive video liveness on the "Age estimation & Matching" demo. 
+* Fix various bugs.
+
 ## July 11, 2025 : Identity Document Autocapture
 
 ### What's New